requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
Faker==25.5.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
validate-docbr==1.10.0
